Update rendering code architecture
Now the logic routes much better between Minecraft and Hero code. It's not perfect, and there is less MC-independent rendering code than I would like, but it is certainly an improvement. Also lots of cleanup to the existing classes (but still more work to be done in that regard).

Also updated the config and fixed the mod name.