Add fade in animation
Basic, for one element, but it is a great proof of concept. 

Also realized that trying to separate out the Minecraft/Hero rendering was too ambitious; it will simply be clean code that is tied together/well documented (it's not happening otherwise, or perhaps that will be a later term refactor).